fixed vector names

+++ src/main/kotlin/gay/j10a1n15/sillygames/data/KeybindSet.kt
@@ -1,7 +1,7 @@
 package gay.j10a1n15.sillygames.data
 
 import gay.j10a1n15.sillygames.config.Config
-import gay.j10a1n15.sillygames.utils.Vector2d
+import gay.j10a1n15.sillygames.utils.vectors.Vector2i
 
 class KeybindSet(
     val keybindUp: Int,
@@ -10,12 +10,12 @@ class KeybindSet(
     val keybindRight: Int,
 ) {
 
-    fun getDirection(key: Int): Vector2d? {
+    fun getDirection(key: Int): Vector2i? {
         return when (key) {
-            keybindUp -> Vector2d(0, -1)
-            keybindDown -> Vector2d(0, 1)
-            keybindLeft -> Vector2d(-1, 0)
-            keybindRight -> Vector2d(1, 0)
+            keybindUp -> Vector2i(0, -1)
+            keybindDown -> Vector2i(0, 1)
+            keybindLeft -> Vector2i(-1, 0)
+            keybindRight -> Vector2i(1, 0)
             else -> null
         }
     }
